requirements.txt
# requirements.txt

python-telegram-bot
python-dotenv